+++ packages/hardhat-core/src/internal/core/providers/construction.ts
@@ -11,14 +11,13 @@ import type {
   ProjectPathsConfig,
 } from "../../../types";
 
-import { HARDHAT_NETWORK_NAME } from "../../constants";
-import { ModulesLogger } from "../../hardhat-network/provider/modules/logger";
-import {
+import type {
   ForkConfig,
   MempoolOrder,
 } from "../../hardhat-network/provider/node-types";
-import { getForkCacheDirPath } from "../../hardhat-network/provider/utils/disk-cache";
-import { BigIntUtils } from "../../util/bigint";
+import type * as ModulesLoggerT from "../../hardhat-network/provider/modules/logger";
+import type * as DiskCacheT from "../../hardhat-network/provider/utils/disk-cache";
+import { HARDHAT_NETWORK_NAME } from "../../constants";
 import { parseDateString } from "../../util/date";
 
 import { normalizeHardhatNetworkAccountsConfig } from "./util";
@@ -72,9 +71,7 @@ export function createProvider(
     ) {
       forkConfig = {
         jsonRpcUrl: hardhatNetConfig.forking?.url,
-        blockNumber: BigIntUtils.mapBigIntToNumber(
-          hardhatNetConfig.forking?.blockNumber
-        ),
+        blockNumber: hardhatNetConfig.forking?.blockNumber,
         httpHeaders: hardhatNetConfig.forking.httpHeaders,
       };
     }
@@ -82,6 +79,11 @@ export function createProvider(
     const accounts = normalizeHardhatNetworkAccountsConfig(
       hardhatNetConfig.accounts
     );
+
+    const { ModulesLogger } =
+      require("../../hardhat-network/provider/modules/logger") as typeof ModulesLoggerT;
+    const { getForkCacheDirPath } =
+      require("../../hardhat-network/provider/utils/disk-cache") as typeof DiskCacheT;
 
     eip1193Provider = new HardhatNetworkProvider(
       hardhatNetConfig.hardfork,