+++ packages/hardhat-core/test/fixture-projects/solidity-config-warnings/unsupported-new-solc.js
@@ -1,3 +1,3 @@
 module.exports = {
-  solidity: "0.8.10",
+  solidity: "0.8.17",
 };