+++ packages/hardhat-core/test/internal/hardhat-network/provider/modules/hardhat.ts
@@ -28,7 +28,7 @@ import {
 import { compileLiteral } from "../../stack-traces/compilation";
 import { getPendingBaseFeePerGas } from "../../helpers/getPendingBaseFeePerGas";
 import { RpcBlockOutput } from "../../../../../src/internal/hardhat-network/provider/output";
-import { BigIntUtils } from "../../../../../src/internal/util/bigint";
+import * as BigIntUtils from "../../../../../src/internal/util/bigint";
 import { EXAMPLE_DIFFICULTY_CONTRACT } from "../../helpers/contracts";
 
 describe("Hardhat module", function () {
@@ -2073,7 +2073,7 @@ describe("Hardhat module", function () {
           await this.provider.send("hardhat_setStorageAt", [
             DEFAULT_ACCOUNTS_ADDRESSES[0],
             numberToRpcQuantity(0),
-            `0x${BigIntUtils.toWord(targetStorageValue)}`,
+            `0x${BigIntUtils.toEvmWord(targetStorageValue)}`,
           ]);
 
           const resultingStorageValue = await this.provider.send(
@@ -2110,7 +2110,7 @@ describe("Hardhat module", function () {
           await this.provider.send("hardhat_setStorageAt", [
             contractAddress,
             numberToRpcQuantity(0),
-            `0x${BigIntUtils.toWord(10n)}`,
+            `0x${BigIntUtils.toEvmWord(10n)}`,
           ]);
 
           // Assert: Verify that the contract retrieves the modified value.
@@ -2169,7 +2169,7 @@ describe("Hardhat module", function () {
           await this.provider.send("hardhat_setStorageAt", [
             DEFAULT_ACCOUNTS_ADDRESSES[0],
             numberToRpcQuantity(0),
-            `0x${BigIntUtils.toWord(targetStorageValue)}`,
+            `0x${BigIntUtils.toEvmWord(targetStorageValue)}`,
           ]);
 
           // Act 2: Mine a block